Pull mutations back into RepositoryStar component
